# refactoring of kmeans + type fixes