With apologies for being back so soon... 

This applies the following changes

"**major**"
1. `x_int = rand(Int16, 10, 20); kmeans(x_int, 3)` ~~is now processed by converting `x_int`  into `Float64` (i.e. something sensible is returned instead of an error)~~ now works by making sure the centers are of an appropriately wide type see #140 for which a similar solution could be applied. Test added for that case.
2. types are now checked for consistency. Effectively at the entry points, the types for `X`, `centers` and `weights` can be whatever the user wants ~~but `_kmeans!` is effectively called on one single type which is `eltype(X)`. Since this happens after conversion, we're sure the result is a floating point number of some type `T` and all operations, distances and updates are done in this type `T`.~~

"**minor**"
- following comments from #138, the style is unified somewhat with identation etc. 
- broadcasting with views is still not applied (I tried a number of ways, all were slower and allocated more) but some simplifications are applied (e.g. the bound check for `1<=assignment<=k` is always verified + in-place computations where appropriate.

I've added comments preceded by "NOTE REFACT:" to justify some of the changes hoping that would help with the reviews, these would be eventually removed of course.

**Mini benchmark**

(same as for #138)

**before**
```julia
julia> using BenchmarkTools, Random, Clustering; Random.seed!(1641); X = randn(500, 5000); w = abs.(randn(5000)); @btime kmeans($X, 3); @btime kmeans($X, 3, weights=w);
  137.222 ms (278 allocations: 1.63 MiB)
  95.003 ms (184 allocations: 1.13 MiB)
```

**after** (992c8bb)
```julia
julia> using BenchmarkTools, Random, Clustering; Random.seed!(1641); X = randn(500, 5000); w = abs.(randn(5000)); @btime kmeans($X, 3); @btime kmeans($X, 3, weights=w);
  106.599 ms (230 allocations: 1.32 MiB)
  83.018 ms (180 allocations: 1.05 MiB)
```