@tlienart 2019-02-05
Ok following comments (thanks for the detailed feedback) here are some modifications (a8d41e4) hopefully going in the right direction.

**major-ish**

1. as per @alyst's comment the convention in Julia with `fun_name!(...)` is not respected by `kmeans!` (mutates second argument). I changed that as it's exported but left the internal functions (that also don't respect the convention but mutate several arguments anyway) untouched.
1. types, I believe we may slowly be moving towards a consensus. In `kmeans!` centers are now forced to be `AbstractFloat` (as per @alyst's suggestion), a warning is raised if the type is not wide enough. In most situations this would not happen as the centers will have generated in `kmeans(...)` and made to have a precisely wide enough type.
1. type naming is simpler in `KmeansResult` as per comments, the only two letter one is `WC` so that it's not mixed with the type of `weights`. 
1. parametrisation of types is implicit in functions unless there are explicitly two arguments 
that must have the same type

**minors**
1. `k::Integer` is now allowed and rounded before the call of `_kmeans...`
1. `init` in `kmeans` is explicitly typed to `::Symbol`
1. `∈` replaced by `in`
1. comment added to specify that `a = argmin(view(....))` is necessarily between `1` and `k` and therefore ok as an index in an `@inbounds` environment
1. basic docstrings added to `kmeans!` and `kmeans`

Ps: apologies for the noob mistake of trying to change too many changes in one PR resulting in big reviews...