@nalimilan 2019-02-09
Note that the docs contain a copy of the `struct` definition which also needs updating.