@alyst 2019-02-10
> Note that the docs contain a copy of the `struct` definition which also needs updating.

Actually, what is the recommended way to organize Julian package documentation nowadays?
It seems a bit suboptimal to document functions and types in the docs, when we can just use the docstrings in-place.
Should we migrate API documentation to docstrings and keep the docs for higher-level things?