@nalimilan 2019-02-10
> Actually, what is the recommended way to organize Julian package documentation nowadays?
> It seems a bit suboptimal to document functions and types in the docs, when we can just use the docstrings in-place.
> Should we migrate API documentation to docstrings and keep the docs for higher-level things?

Yes. Ideally docs would be ported from RST to Documenter, and docstrings automatically inserted into the manual.