@tlienart 2019-02-12
Thanks a lot for the reviews. It seems the two remaining sticking points are:

1. in `kmeans!` whether to check for the type of the update to match that of given `centers`  and what to do if it doesn't match (I suggest warning, or error)
2. the result `struct`, whether to parametrize `cost` and `cweight` or not, if not I suggest at least not casting things to Float64 unnecessarily.

Small PS: happy to modify the docs following recommendations.