@tlienart 2019-02-15
Bump? I believe I've either removed or modified all contentious bits of code :)