@tlienart 2019-02-21
Thanks for the reviews @alyst , I think (hope) I've applied all your suggestions