@alyst 2019-03-04
@tlienart Could you please add the two tiny changes I've noted? Besides that, I think it's ready to be merged.
@nalimilan There's #147, which is conflicting with this PR, but this one was created earlier and came through so many iterations that I think it would be fair if it is merged first.