@nalimilan 2019-03-05
> @nalimilan There's #147, which is conflicting with this PR, but this one was created earlier and came through so many iterations that I think it would be fair if it is merged first.

Fine, anyhow rebasing shouldn't be hard. But CI currently fails for this PR.