@tlienart 2019-03-05
The failure is unrelated to the PR, I think it’s due to changes in Distances, #147 also fails on the same I think.

What do you want me to do? I can try merging the changes of 147 here along with changes to fix the failure related to distances but that will protract the merge..