@nalimilan 2019-03-05
OK, since what you did is essentially a rebased, I've merged #147 and finished the rebasing. It would have been weird to merge this PR first with changes from #147 which are completely unrelated to this PR.