@alyst 2019-03-12
@tlienart Sorry for the delay; finally merged! :) Thanks for your hard work! Now that we've converged on the requirements for the clustering methods refactoring, you are welcome to "refresh" the other ones. I think it brings the code base up to the current Julia standards and improves the maintainability in the long run.